Optimize when FilterImpl is used for service reference lookup
It is common for a service lookup to be done with no supplied
class name.  This is common for ServiceTracker when a Filter
object is used during construction.  It is also how
the Felix SCR implementation looksup services when determining
if the requirements are met for a service component.

This fix uses the FilterImpl method getRequiredObjectClass
to try and determine the class name a service filter is
trying to lookup so it can drastically reduce the number
of service references that need to be evaluated against
the supplied filter.