Fix issue #52 - handle registered and dependent policy for listResources
When the OSGi API BundleWiring was added a new way to search
for resources was added with the listResources method.  The
buddy loading policy was not considered with the Equinox
implementation of this new listResources method.

This address that for only the registered and dependent policies.
Other policies involve searching the system class loaders, which
we cannot easily list all resources, or the global policy,
which searches all exported packages and would lead to a
very expensive listing of all bundles.

This implementation will only search buddies when the
LISTRESOURCES_LOCAL is not used.